Reinvent another way to get the entry path from the CVS rlog When a remote CVS repository is made of several modules, the `cvs rlog` output shows the real path on the server of each entry, not one relative to the working copy (reasonable, since with 'rlog' you do not have a working copy...).
The previous solution assumed a common prefix for any single
repository, but that turned out to be wrong.

The only way to solve this was to read also cvs rlog stderr output,
where cvs prints something like::

    cvs rlog: Logging Zope/lib/python/DateTime

The parser look for these messages, and applies the right path to
succedings entries.