First stab at fixing deadlock I ran into this trying to commit a very huge changeset with a very large log message to darcs.
If both output and input were specified to the given command, the shwrap.py
would try to write the entire input before reading any output.  This worked
with small inputs and outputs because pipes have buffers.  But it didn't work
with this one, because both buffers filled.  The result: deadlock.

One easy solution is to use threading; shove the output and input sides into
separate threads and run from there.  Works well.