Revised monotone support, merging with Riccardo patch My --monotone_db and --monotone_branch options disappeared. The monotone setup command now has to be called by the user. What I don't like about this is: the user either
a) has to start tailor.py -b -t monotone ... first, get an error
   message, invoke monotone setup in the correct directory and then
   start tailor.py again. Or

b) he creates the directories which tailor creates himself and prepares
   monotone setup there, then launches tailor.py.

With a) you have to cvs update && cvs rlog twice (in case of a CVS
source) which takes a lot of time and networking bandwith. It's obvious
that this isn't a nice way to go. b) requires a lot of knowledge from
the user. This could be documented, but options for tailor would be
nicer, IMHO.

Then there was the problem of monotone beeing called outside of it's
working copy for the very first commit during bootstrap. Riccardo has
solved that by not touching the MT/log but adding a --message options to
monotone. That's nice, too. I've come up with another solution: instead
of monotone commit $SUBDIR we can change to $SUBDIR and issue a plain
monotone commit there. So if MonotoneCommit misses the MT directory it
now checks if only one entry (the subdir) was given. In that case it
changes the working directory to that subdir and commits there. That
solution now works with monotone-0.19 and seems better to me (we never
call monotone outside its working copy).