git commit fixes (date, # at end of commit message)
As of at least git 0.99.9e, git refuses to parse timestamps with
microseconds such as str(datetime) may emit. This patch uses strftime
to force the format to something simpler.

It also manually duplicates GIT_AUTHOR_* to GIT_COMMITTER_* so that
cg-log produces output most similar to the source log.

git-commit expects the commit message to end with a newline, so this
patch adds one if it isn't already there.