Handle old-style darcs date format
Older darcs patches appear to keep the date in the form
Sun Oct 20 20:10:04 EDT 2002 rather than the current
20021020201004. This patch tries to parse the date as the old format
if it fails to read it as the new. NOTE: This older patch format also
calculates the hash from the date in the recorder's time zone, but
darcs pull uses the puller's time zone, so it may not calculate the
correct hash of the patch. Perhaps it's safer to just throw an
exception if the date isn't in the current format?