Fix tkt 17 (let svn do mv itself instead of add/remove)
The problem described in ticket 17 arises because upstream has already
relocated the renamed directory, including the .svn metadir
inside. This metadata directory causes svn to refuse to add a path
that it believes is already under version control. But that metadir
now disagrees with the working directory about where it exists in the
file system.

We have a couple of choices: either put the directory back if possible
and let svn do the merge on its own, or remove the .svn entry from the
new directory and all its subdirectories. I like the first choice, not
just because it's a little easier, but also because it preserves file
history. The current implementation notes that svn mv is equal to svn
cp+rm, but doesn't actually do cp+rm. Instead it does rm+add, which
isn't equivalent.

I had to add --force to the mv operation to handle the case where the
renamed entry or content underneath it had been changed in the same commit.