Ignore comment lines in git "nothing to commit" messages
If there are any unversioned, unignored files in the working directory
git will currently list them in comment lines during commit. Since git
doesn't support empty commits, tailor checks for a 'nothing to commit'
message on failure and treats it as success. This check was broken when
there were comments. Perhaps a far simpler method would be to simply not
perform the commit when entries == [], but I preserve the status quo here.