Use git-rev-list to read tags
The previous tag reading code only understood lightweight tags. This
uses git-rev-list to read real tag objects as well. It might be a good
idea to cache tag info (based on a stat of the file in refs/tags) per
run, to avoid recomputing the same information at every changeset.