Query hg for default options
hg looks up command options from the argument without trapping
KeyError. So whenever an option is added to an hg command, tailor
backtraces until it's been taught to pass in the new option. This
patch queries hg for the set of options for all commands, which
should provide some protection against this particular problem.