Synthesize missing parent directories when adding files to CVS
This does for adds what target._renameEntries does for renames. This code
is only for the CVS module; perhaps it should be relocated to target. But
I don't understand why target is doing the parent creation anyway - for
instance it is unnecessary for git and hg. They could override, but there's
also some replace logic in that routine that they'd have to reimplement.