[bzr] _addSubtree(): do not add files in IGNORED_METADIRS and .bzrignore
The implementation of _addSubtree() recently introduced forces the
addition of files ignored by bzr, because the source repository could
contain files that matched bzr's idea of ignorable files.

However, dirs in IGNORED_METADIRS and things like the status and log
files (which _prepareTargetRepository adds to .bzrignore) should always
be ignored. This patch addresses that.