[bzr] ignore state and log files without using .bzrignore
This patch makes BzrWorkingDir not use .bzrignore to ignore the state
and log files when necessary, and add them to bzrlib's DEFAULT_IGNORE
list.

The rationale for this is that the conversion is cleaner this way, since
otherwise the resulting bzr repository will always need an extra commit
to revert .bzrignore to match the original repository.