Specified peg revision on initial subversion checkout
When attempting to get the path record, subversion needs to know
the exact file you need.  It assumes the HEAD revision contains the
necessary file, but if the directory or file has been deleted,
then the checkout will fail.  If another directory or file has taken
its place in the path, it will checkout the wrong file!  Specifying
the peg version using the "at notation" fixes this problem.