svn-commit-all-files.patch Subversion: new option commit-all-files
If it's true (default), then commit all files with a command "svn commit ."
If it's false, commit list of changed files instead the '.'. This can be
splitt into multiple commits on big tree with many file changes. Than the
first commit was not propset with author and date.
(see MAX_CMDLINE_LENGTH = 8000)