svn-rename-directory-hacking3.patch Ticket #65, #125: Rename direcory fails, because exist .svn directory
If the target reposity has files in subdirectory, then remove the
complete dir before rename. But keep the dir '.svn' or '_CVS'

Hack for the hack:
Remove back all directories '.svn' from the old to new directory.
Remember: TAILOR-HACKED-TEMP-NAME contains the new files!

Please leave the comment graph some there in the source. Was hard to
finding out the idea behind this hack.

FIXME:
o This is only a warkarround for Subversion and should be generaly
  rewriten for all repositories generaly.
  Use METADIR should work for all reposities now. Only tested Subversion
  and Monotone.
o The UTF-8 support for filenames can be broken, because len of filename is
  not the real stringlength. The line
  join(absnew + '-TAILOR-HACKED-TEMP-NAME' + root[len(absnew):], ...
  should written by path splitting and re-join the path.
o shared_basedirs are untouched, not tested and would have same problems.