[Tailor] patch-name-format empty was not working
Changeset [1359] has not fixed it.
Please don't add the revision number, if the user wand to leave empty it.

Examples for no patchname:
    patch-name-format = 
    patch-name-format = ""
The empty string "" is more understandable as a blank space after the '='.

This example is more for internal interestings, it set default patchname:
    patch-name-format = None

All other values, set user specifics.

Tested for Monotone and Subversion. All others have similar code style,
so that should work in all repositories.