git_escape_percent_sign
ExternalCommand._execute() supports %-formatting in the command.
This feature is not used by git repository backend, however
if file name (or other part of a command) contains % sign
an exception will be thrown during %-formatting in
ExternalCommand._execute(). To prevent this, escape (i.e. double)
all percent signs in a command.